widget - logout url with confirm modal
